backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-limiter==3.9.0
flask-talisman==1.1.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.27.2
gunicorn==22.0.0
flask-jwt-extended==4.7.1
google-auth==2.29.0
requests==2.32.3
